fix text offset issues caused by update
getting ready for uni, no time to examine what exactly caused the issue, seems like it was that ``` UiText("", true)``` might no longer apply the proper vertical offset for the font size anymore, which is cringe, and if the new way to use that is horrendous and will make the thing look awful, because of uneven spacing. The menu was also affected on opengl, so the devs are just breaking shiet to break shiet, no other reason. Might have to check if simplerun is affect and fix that abomination too.
Thanks https://steamcommunity.com/profiles/76561198276715031 Forkdude, for reporting the issue.