fix keybind offset
In 1.5.1 update, the game game how it calculates font size. Which is a pain for me, cuz tearwares legacy menu is using this bs for automatic offsetting of menu elements. Now I have to hardcore all of this manually, so I can't just swap a font and be done. Great. I'm pretty sure the old way it would get the font size was more accurate, but who cares.